Separating TimelineItem from main Timeline component
