+++ components/Timeline.js
@@ -1,10 +1,10 @@
 "use client";
 import React, { useEffect, useRef } from 'react';
-import Image from 'next/image';
 import { HiRocketLaunch } from "react-icons/hi2";
 import BarkerLogo from '../public/images/timeline/barker_logo.png';
 import UnswLogo from '../public/images/timeline/unsw_logo.png';
 import InfotrackLogo from '../public/images/timeline/infotrack_logo.png';
+import TimelineItem from './TimelineItem';
 import { motion, useInView, useAnimation } from "framer-motion";
 
 const experiences = [
@@ -41,37 +41,6 @@ const experiences = [
     scale: 0.8
   }
 ];
-
-function TimelineItem({ data }) {
-  return (
-    <div className="timeline-item flex space-x-4 my-8" >
-      <div className="-ml-2 mt-8 w-4 h-4 bg-default-600 rounded shadow-xl" />
-      <div
-        className="w-20 h-20 rounded-xl relative shadow-xl flex justify-center items-center"
-        style={{ backgroundColor: data.backgroundColor }}
-      >
-        {data.image && (
-          <Image
-            src={data.image}
-            style={{ objectFit: "contain", transform: `scale(${data.scale})` }}
-            fill={true}
-            alt={"Logo for " + data.name}
-            sizes="160px"
-          />
-        )}
-        {data.icon}
-      </div>
-      <div className="flex-1">
-        <div>
-          <h3 className="text-xl mb-2">{data.type}</h3>
-          <h3 className="text-xl font-bold mb-2">{data.name}</h3>
-          <p className="text-gray-500">{data.duration}</p>
-          <p className="mt-2 text-gray-700">{data.description}</p>
-        </div>
-      </div>
-    </div>
-  );
-}
 
 function Timeline() {
   const ref = useRef(null);
